Extract work info panel
